extract yaml::ToOpenApi from gesha-core
