+++ libs/gesha-core/src/conversions/error.rs
@@ -1,19 +1,13 @@
-use crate::conversions::Error::Yaml;
-use crate::yaml;
 use openapi_types::v3_0::OpenApiDataType;
 
 pub type Result<A> = std::result::Result<A, Error>;
 
 #[derive(Debug)]
 pub enum Error {
-    // inherited errors
-    Yaml(yaml::Error),
-
     // client errors
     ReferenceObjectNotFound(String),
 
     // module errors
-    IncompatibleVersion,
     TransformBroken {
         detail: String,
     },
@@ -21,13 +15,6 @@ pub enum Error {
         data_type: OpenApiDataType,
         format: String,
     },
-    UnknownDataType(String),
-}
-
-impl From<yaml::Error> for Error {
-    fn from(cause: yaml::Error) -> Self {
-        Yaml(cause)
-    }
 }
 
 #[macro_export]
