+++ libs/gesha-core/src/conversions/v3_0/mod.rs
@@ -1,2 +1,1 @@
-mod to_openapi;
 mod to_rust_type;
