+++ libs/gesha-core/src/gateway/reader.rs
@@ -1,8 +1,7 @@
-use crate::conversions::ToOpenApi;
 use crate::conversions::ToRustType;
 use crate::gateway::Error::CannotReadFile;
 use crate::gateway::{Error, Result};
-use crate::yaml::{load_from_str, YamlMap};
+use openapi_types::yaml::{load_from_str, ToOpenApi, YamlMap};
 use std::fmt::Debug;
 use std::fs;
 use std::marker::PhantomData;
@@ -30,7 +29,7 @@ where
     {
         let path = path.as_ref();
         let map = open_yaml_map(path)?;
-        let openapi_value = ToOpenApi::apply(map).map_err(Error::conversion(path))?;
+        let openapi_value = ToOpenApi::apply(map).map_err(Error::openapi(path))?;
         let rust_type = ToRustType::apply(openapi_value).map_err(Error::conversion(path))?;
         Ok(rust_type)
     }
@@ -45,7 +44,7 @@ pub fn file_to_string<A: AsRef<Path>>(path: A) -> Result<String> {
 }
 
 fn open_yaml_map<A: AsRef<Path>>(path: A) -> Result<YamlMap> {
-    let content = file_to_string(path)?;
-    let map = load_from_str(&content)?;
+    let content = file_to_string(path.as_ref())?;
+    let map = load_from_str(&content).map_err(Error::openapi(path.as_ref()))?;
     Ok(map)
 }
