+++ libs/gesha-core/src/lib.rs
@@ -4,4 +4,3 @@ pub mod trace;
 mod conversions;
 mod misc;
 mod renderer;
-mod yaml;
