+++ libs/openapi-types/src/lib.rs
@@ -1,1 +1,5 @@
 pub mod v3_0;
+pub mod yaml;
+
+mod error;
+pub use error::{Error, Result};
