+++ libs/openapi-types/src/v3_0/document.rs
@@ -1,0 +1,17 @@
+use crate::v3_0::{ComponentsObject, PathsObject};
+
+/// OpenAPI Document
+/// https://github.com/OAI/OpenAPI-Specification/blob/main/versions/3.0.3.md#schema
+#[derive(Debug)]
+pub struct Document {
+    pub openapi: String,
+    pub info: InfoObject,
+    pub paths: PathsObject,
+    pub components: Option<ComponentsObject>,
+}
+
+/// https://github.com/OAI/OpenAPI-Specification/blob/main/versions/3.0.3.md#infoObject
+#[derive(Debug)]
+pub struct InfoObject {
+    pub title: String,
+}
