+++ libs/openapi-types/src/yaml/to_openapi.rs
@@ -1,0 +1,7 @@
+use crate::yaml::YamlMap;
+use crate::Result;
+
+/// convert YamlMap to a type defined in this crate.
+pub trait ToOpenApi: Sized {
+    fn apply(map: YamlMap) -> Result<Self>;
+}
